Assignment 6: Coffee Shop Line
Problem Statement -
Arrival: Customers arrive at the coffee shop and stand in line. 
Order Processing: The first customer in line gets their order taken, and the barista starts making the coffee.
Serving: Once the first customer is served, they leave the queue, and the next customer in line moves forward to be served. 
Write a program to implement a simple queue